requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.0
nltk>=3.8.0
numpy>=1.21.0